checking before delete Nhanvien and getAll function Khachhang

+++ src/main/java/com/herokuapp/realestatebk/service/NhanvienService.java
@@ -1,12 +1,14 @@
 package com.herokuapp.realestatebk.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.herokuapp.realestatebk.entity.Nhanvien;
 import com.herokuapp.realestatebk.form.FormLogin;
+import com.herokuapp.realestatebk.repository.KhachhangRespository;
 import com.herokuapp.realestatebk.repository.NhanvienRepository;
 
 
@@ -15,13 +17,25 @@ public class NhanvienService {
 	
 	@Autowired
 	private NhanvienRepository nhanvienRepository;
+	
+	@Autowired
+	private KhachhangRespository khachhangRespository;
 
 	public List<Nhanvien> getAllNhanvien() {
 		return (List<Nhanvien>) nhanvienRepository.findAll();
 	}
 
-	public void deleteNhanvien(int id) {
-		nhanvienRepository.deleteById(id);;
+	public String deleteNhanvien(int id) {
+		int checkKhachhang = khachhangRespository.getKhachhangByNvID(id);
+		if(checkKhachhang == 0) {
+			nhanvienRepository.deleteById(id);
+			return "Deleted Successfully";
+		}else{
+		Optional<Nhanvien> nhanvien = nhanvienRepository.findById(id);
+		nhanvien.get().setTrangthai(Byte.valueOf("1")); 
+		nhanvienRepository.save(nhanvien.get());
+		return "Disabled Account Successfully";
+		}
 	}
 
 	public Nhanvien addNhanvien(Nhanvien nhanvien) {
@@ -36,16 +50,6 @@ public class NhanvienService {
 			nhanvienEdit = nhanvienRepository.save(nhanvien);
 		}
 		return nhanvienEdit;
-	}
-	
-	public Nhanvien checkExistNhanvien(String email, String pass) {
-		Nhanvien checkedNhanvien = null;
-		List<Nhanvien> listNhanvien = (List<Nhanvien>) nhanvienRepository.findAll();
-		for (Nhanvien nv : listNhanvien) {
-			if(email.equals(nv.getEmail()) && pass.equals(nv.getMatkhau()))
-				checkedNhanvien = nv;
-		}
-		return checkedNhanvien;
 	}
 
 	public Nhanvien login(FormLogin formLogin) {
